% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omuecon1a-package.R
\docType{package}
\name{omuecon1a-package}
\alias{omuecon1a}
\alias{omuecon1a-package}
\title{omuecon1a: Seminar Matching App for the Department of Economics OMU}
\description{
This is probably not what you might want to use!
}
\author{
\strong{Maintainer}: Kenji Sato \email{kenji@kenjisato.jp} (\href{https://orcid.org/0009-0000-4520-2560}{ORCID}) [copyright holder]

}
\keyword{internal}
