% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Prepare project directory}
\usage{
prep(dir = format(Sys.Date(), "matching\%y"), parent = NULL)
}
\arguments{
\item{dir}{character. The name of the directory to extract sample files.}

\item{parent}{character. The name of the enclosing directory. Directory
picker will open when unspecified (or NULL).}
}
\value{
Invisibly returns the path to the project directory.
}
\description{
Extract sample files, the recommended structure for the matching project.
This function is intended to be used in the interactive shell.
}
