% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{up}
\alias{up}
\title{Launch shiny app}
\usage{
up(config = NULL)
}
\arguments{
\item{config}{character. path to config file, read by `read_config()`.}
}
\value{
Launch the shiny app.
}
\description{
Launch shiny app
}
